chore: move and rename handleAdditionalTextEdits method
